server/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pillow
python-dotenv
paddlepaddle
paddleocr
google-generativeai
pytesseract
